Fix total count issue

+++ messaging-service/src/main/kotlin/com/forge/messageservice/graphql/extensions/Connection.kt
@@ -4,6 +4,6 @@ import graphql.relay.Connection
 
 interface Connection<T> : Connection<T>{
 
-    fun totalCount(): Int
+    fun totalCount(): Long
     
 }
